feat: Simple SMTP implementation with easy client and message building
